....I..... [ZBXNEXT-801] fix compilation errors on Windows
git-svn-id: svn://svn.zabbix.com/branches/dev@28042 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-801/src/libs/zbxcommon/str.c
@@ -1911,7 +1911,7 @@ size_t	zbx_binary2hex(const u_char *input, size_t ilen, char **output, size_t *o
 	}
 	o = *output;
 
-	while (i - input < ilen)
+	while ((size_t)(i - input) < ilen)
 	{
 		*o++ = zbx_num2hex((*i >> 4) & 0xf);
 		*o++ = zbx_num2hex(*i & 0xf);
